Update OpenAPI spec (#160)
## Changes
Update of the OpenAPI spec. This also introduces the template_libraries
field being added in
https://github.com/databricks/databricks-sdk-go/pull/635.

This PR contains some other changes:
1. Change `float`/`Float` to `double`/`Double`. OpenAPI specs currently
specify `format: "double"` for all non-integral numeric values.
2. Remove old types that were previously in the OpenAPI spec but are no
longer needed.
3. Do not generate files for types that don't correspond to distinct
Java types (e.g. are not objects or enums). A lot of extra files are
generated unnecessarily today, corresponding to either
integers/strings/booleans or arrays/maps. Since Java has no type-def
functionality, these are represented in the SDK as their JDK
equivalents.

## Tests
Added AccountMetastoreAssignmentsIT to verify the correct functionality
of the updated AccountMetastoreAssignments API.
Nightly tests passed on this PR.

+++ databricks-sdk-java/src/main/java/com/databricks/sdk/mixin/ClustersExt.java
@@ -165,7 +165,7 @@ public class ClustersExt extends ClustersAPI {
   private final Comparator<NodeType> nodeSortingComparator =
       (item1, item2) ->
           Comparator.comparing(NodeType::getIsDeprecated, Comparator.nullsLast(Boolean::compare))
-              .thenComparing(NodeType::getNumCores, Comparator.nullsLast(Float::compare))
+              .thenComparing(NodeType::getNumCores, Comparator.nullsLast(Double::compare))
               .thenComparing(NodeType::getMemoryMb, Comparator.nullsLast(Long::compare))
               .thenComparing(
                   instanceTypeComparator(NodeInstanceType::getLocalDisks),
