+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/ml/ExperimentsAPI.java
@@ -309,7 +309,7 @@ public class ExperimentsAPI {
     impl.logInputs(request);
   }
 
-  public void logMetric(String key, float value, long timestamp) {
+  public void logMetric(String key, double value, long timestamp) {
     logMetric(new LogMetric().setKey(key).setValue(value).setTimestamp(timestamp));
   }
